// src/app/ws/soccer/model/CommentDTO.ts
/**
 * Soccer API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * Contact: voetbalsvk@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import * as models from "./models";

export interface CommentDTO {
    content?: string;

    editable?: boolean;

    id?: number;

    postDate?: string;

    postedBy?: models.AccountDTO;

}